pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "photocluster"
version = "0.1.0"
description = "Scan, cluster, and organise photos by location and time"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "typer>=0.9.0",
    "textual>=0.50.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
    "piexif>=1.1.3",
    "geopy>=2.4.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
photocluster = "photocluster.main:app"

[tool.hatch.build.targets.wheel]
packages = ["photocluster"]

[tool.pytest.ini_options]
testpaths = ["tests"]
